🚨 make the for loop variable size_t
For the following statements

for (int i = 0; i < _qubits.size(); ++i)

where _qubits is a vector, the compiler will complain about the
dismatched type of i and _qubits.size(). I change the type of i
to size_t.

+++ src/qcCheck.cpp
@@ -289,7 +289,7 @@ void Checker::addElementToOutputJSON(const std::string key,
 void Checker::printOutputJSON() const
 {
     std::cout << "{\n";
-    for (int i = 0; i < _outputJSON.size(); ++i)
+    for (size_t i = 0; i < _outputJSON.size(); ++i)
     {
         auto &p = _outputJSON[i];
         std::cout << "  ";
